+++ tomcat/src/main/java/org/apache/coyote/http/Constants.java
@@ -2,10 +2,10 @@ package org.apache.coyote.http;
 
 public class Constants {
 
-    static final String SPACE = " ";
-    static final String CRLF = "\r\n";
-    static final String COLON = ":";
-    static final String COLON_WITH_SPACE = ": ";
+    public static final String SPACE = " ";
+    public static final String CRLF = "\r\n";
+    public static final String COLON = ":";
+    public static final String COLON_WITH_SPACE = ": ";
 
     private Constants() {}
 }
