+++ tomcat/src/main/java/org/apache/coyote/http/request/Path.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.request;
 
 public class Path {
