+++ tomcat/src/main/java/org/apache/coyote/http/request/RequestLine.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.request;
+
+import org.apache.coyote.http.HttpVersion;
 
 import java.util.List;
