+++ tomcat/src/main/java/org/apache/coyote/http/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.response;
 
 import static org.apache.coyote.http.Constants.CRLF;
