+++ tomcat/src/main/java/org/apache/coyote/http/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.response;
 
 public enum HttpStatus {
     OK("OK", 200),
