+++ tomcat/src/test/java/org/apache/coyote/http/RequestLineTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http;
 
+import org.apache.coyote.http.request.RequestLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
